New Navigation Bar and Ongoing Home Page
The new navigation has its own logo and working links, with color theme of orange and white. The nav will be also always on top so the user wont scroll up to reach it. Also add some unique animations to improve the site quality.  This nav is not done ye.