Simple And Finished Website
Website is now working, with some removals and additions. About Us is updated, Projects is changed from HTML Facebook Page. Buttons is change to links. Changes in services. ChatBox is soon to implemented but not now so we're deleting it for now. Changes in footer. Contact Us is now working with backend api.